+++ ATS_API/Scripts/Helpers/TextureHelper.cs
@@ -33,6 +33,7 @@ public static class TextureHelper
         GoodIcon = 8,
         BiomeIcon = 9,
         BiomeWorldMapTexture = 10,
+        DecorationTierIcon = 11,
     }
 
     private static Vector2 DEFAULT_PIVOT = new(0.5f, 0.5f);
@@ -50,6 +51,7 @@ public static class TextureHelper
         { SpriteType.GoodIcon, new Rect(0.0f, 0.0f, 128.0f, 128.0f) },
         { SpriteType.BiomeIcon, new Rect(0.0f, 0.0f, 128.0f, 128.0f) },
         { SpriteType.BiomeWorldMapTexture, new Rect(0.0f, 0.0f, 2048.0f, 2048.0f) },
+        { SpriteType.DecorationTierIcon, new Rect(0.0f, 0.0f, 128.0f, 128.0f) },
     };
 
     private static readonly Dictionary<SpriteType, Vector2> SPRITE_PIVOTS = new()
@@ -65,6 +67,7 @@ public static class TextureHelper
         { SpriteType.GoodIcon, DEFAULT_PIVOT },
         { SpriteType.BiomeIcon, DEFAULT_PIVOT },
         { SpriteType.BiomeWorldMapTexture, DEFAULT_PIVOT },
+        { SpriteType.DecorationTierIcon, DEFAULT_PIVOT },
     };
     
     public static Texture2D GetWhiteTexture(SpriteType spriteType)